chore(repo-hygiene): move remaining root utilities to devtools with shims
